Add files via upload
I looked up how to change the file name of my project as I had named it "Beatles" following along in lecture. Also, my slider bar does not work when I try running the app but shows up in previews in Xcode. I am not sure if this is because the remote computers updated to the new version of iOS or not.